refactor data.ts
