+++ frontend/public/data/socialInfos.ts
@@ -1,0 +1,43 @@
+export type socialInfo = {
+  href: string;
+  src: string;
+  alt: string;
+};
+  
+export const socialLinks: socialInfo[] = [
+  {
+    href: 'https://bit.ly/CSESocDiscord',
+    src: '/assets/discord_logo.svg',
+    alt: 'Discord'
+  },
+  {
+    href: 'https://www.facebook.com/csesoc/',
+    src: '/assets/fb_logo.svg',
+    alt: 'Facebook'
+  },
+  {
+    href: 'https://twitter.com/csesoc?lang=en',
+    src: '/assets/x_twitter_icon.svg',
+    alt: 'X/Twitter'
+  },
+  {
+    href: 'https://www.youtube.com/@CSESocUNSW',
+    src: '/assets/youtube_logo.svg',
+    alt: 'Youtube'
+  },
+  {
+    href: 'https://www.instagram.com/csesoc_unsw/?hl=en',
+    src: '/assets/instagram_logo.svg',
+    alt: 'Instagram'
+  },
+  {
+    href: 'https://www.tiktok.com/@csesoc?lang=en',
+    src: '/assets/tiktok_logo.svg',
+    alt: 'TikTok'
+  },
+  {
+    href: 'https://www.linkedin.com/company/csesoc?originalSubdomain=au',
+    src: '/assets/linkedin_logo.svg',
+    alt: 'LinkedIn'
+  }
+];
